refactor(cypress): move query-manager fns out of datasource into querymanager
The query PANEL/manager operations were living in the datasource file. Move them
to appBuilder/querymanager/queries.js and dedup:
- moved: verifypreview, addQueryN, addQuery, addDsAndAddQuery, addQueryAndOpenEditor,
  createDataQuery, createRestAPIQuery, verifyPreviewData
- deduped: removed dataSource's `query` (identical copy already in queries.js);
  also dropped a duplicate `url:` key in createRestAPIQuery
- dataSource.js now holds datasource-CONNECTION helpers only
All importers repointed (single-name path swaps + one bundled split in
dataSourcePermissions); 0 dangling; webpack smoke test green. Comments preserved.
