+++ cypress-tests/cypress/e2e/happyPath/platform/ceTestcases/dataSources/dataSourcePermissions.cy.skip.js
@@ -7,7 +7,6 @@ import {
 import { commonText } from "Texts/common";
 import {
   deleteDatasource,
-  addQuery,
   verifyValueOnInspector,
 } from "Support/utils/marketplace/datasources/dataSource";
 import { dataSourceSelector } from "Selectors/dataSource";
@@ -19,7 +18,7 @@ import {
 } from "Support/utils/common";
 import { inviteUserBasedOnRole } from "Support/utils/manageGroups";
 import { resolveHost } from "Support/utils/apps";
-import { addSuccessNotification } from "Support/utils/appBuilder/querymanager/queries";
+import { addQuery, addSuccessNotification } from "Support/utils/appBuilder/querymanager/queries";
 
 const data = {};
 data.firstName = fake.firstName.toLowerCase().replaceAll("[^A-Za-z]", "");
